chore: Update progress_to_next_level field in MeditationProgressSerializer

+++ thoughts_app_service/thoughts_app/thoughts_core/serializers.py
@@ -116,4 +116,4 @@ class UserRegistrationSerializer(serializers.ModelSerializer):
 
 class MeditationProgressSerializer(serializers.Serializer):
     level_name = serializers.CharField()
-    progress_to_next_level = serializers.FloatField()
+    progress_to_next_level = serializers.DecimalField(max_digits=3, decimal_places=2)
